updates to connection, added whitelists, fixed blacklists and records
Added handling to the connection class to handle TypeErrors due to 404s
returning an error string instead of object. Also changed the
connection so that, rather than the json body alone, it now returns an
object with 'body' and 'headers' keys. POST requests return a
'location' header with the uri of the created service. Renamed 'items'
to 'records' to match the uri. Fixed blacklists create, non-admins do
not need to supply a sponsor ID. Added whitelists method.